bots/vk/bot1_make/requirements1.txt
aiogram==3.3.0
gspread==5.11.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
python-dotenv==1.0.0
redis>=5.0.0
vk-api==11.9.9
peewee==3.16.2